A thing, though safe interface is actually unsafe
e.g, add_listener returns dangiling pointer. I mean, it is an unsafe
interface which is correct, but it should be more idiot proof than that.

E.g the InputManager should contain its' inner data in a box already, so
that I can use it and not worry about it moving around and creating
dangling pointers everywhere.