Added ability to terminate compositor
Note I'm not happy with this necessarily, but we can't give _any_ access
to the compositor struct after the loop has started because the
compositor struct holds the entire program state (and thus it would
break all of the aliasing rules).

This is _very_ limited form of it, and because we store it in an
UnsafeCell there's no guarantees about aliasing so we are sure this will
work.

In the future this may be revised. Currently I don't think any state
will be stored in Compositor, but it's possible that is a mistake.
The current lazy_static way of Way Cooler has its problems even if it
works, but it's possible I could make something better here.