choose_best_mode no longer exposed, is unsafe
You can do whatever action you want on the output when it's removed, and
setting the mode results in UB. So, the only way to set the mode is in
the OutputBuilder during creation. After that you cannot change it.

This restriction is intense, and can be lifted in the future if you want
mode switching at runtime. This is the best option that solves both
issues, and in the future it can be "fixed" by re-exposing the function,
making it safe by restricting the actions that can be done on an output
during destruction by putting it in a restricted wrapper type.