PointerHandle -> Pointer. Made a dif PointerHandle
The handle now functions like it does in wlc.rs. I.E: it allows users to
refer to a Pointer but still have it wholly owned by wlroots/compositor
user.

One limitation that's not present in wlc.rs: You _cannot_ downgrade a
previously upgraded PointerHandle (or any future handles added to
wlroots-rs).

This is due to not wanting to pay the cost of keeping a list of live
values in a lazy_static.

Right now we also use Rc, not Arc. This is because adding thread safety
is an additive feature, and not one I'm sure I can do yet. Later this
might change, but right now it's very, very cheap to use an Rc so I'm
going to stick with this.