Output and OutputHandle
Had to change how it's stored internally, but I think this works better.

I need to ensure that the way you destroy outputs doesn't conflict with
this memory model. E.g if you destroy an output, does it immediantly
destroy it or is it queued up in wlroots?

I assume the latter, but if it's the former than I need to rethink some
of this / give up and make destruction unsafe. That will be ok though,
because the unsafety is limited -- e.g if you destroy it, you gotta just
not use that handle again within that callback, which is easy enough to
audit for.

We shall see. Depends on the relationship between OutputLayout and
Output. Which will probably change now that OutputLayout can just store
an OutputHandle which will mean no more ugly Rc<Refcell<OutputLayout>>
bullshit in Output.