Add panic catchers so we don't unwind into C
The way it works is that we add it to the macro so that all callbacks
from the handler traits will have to go through a panic catcher that
will stop it from unwinding back into wlroots/wayland.

Once we catch it, we then terminate the compositor and let the rest of
the code run (which won't be much, since we called wl_display_terminate
it will stop the loop getting events and just let the rest of the code,
of which there is 0 in the macro, run to completion and pop back up the
stack back into C land).

Before terminating the compositor, we store the error in the Compositor
structure so that we can re-throw it later.

Once it's terminated, in our mainloop function we check the internal
error conditional that's set when a panic is thrown (see above). If it's
set, then we re-throw using std::panic::resume_unwind.