Don't panic we find an output that's not ours
An output always spawns if we are running in an embedded environment
(e.g while under an already running Wayland instance). This for example
causes the 'minimal' example to break badly when closing the window
otherwise.

This isn't a big deal, we just see if we handle it and if we don't we
just keep going.