Plug safety hole from multi mut handles
This plugs the hole that can be caused from running multiple handles
within the `run` callback that all point to the same backing wlroots
owned storage.

It was probably safe, tbh, because internally they all use *mut to get
their data and it probably syncs that up just fine (e.g there's no
incorrect way for Rust to re-order / elide the data fetching because
it's behind an opaque barrier). However, UB is UB, so this limitation
has been added.

The solution is to add an AtomicBool to the Rc in the concrete types and
check that each time we acquire it. This means we can no longer expose
the unsafe `upgrade` function, but that's not a problem since that
wasn't supposed to be used anyways. The reason we can't expose it is
that there's no way to set the bool back to false after it's done
executing since once you upgrade you lose that information (because we
can't let the Rc live longer than 1, this is an Rc but really it's just
a bunch of weak pointer to one owner).