{Output,Keyboard,Pointer}Handle.run unwind safe
If the run function panics, we clean up the state of the handle (e.g set
it to false). If the function that panicked somehow messed up the
internal state of the handle we abort (by panicking after a panic).

The main use case of this is if the user has a panic catcher around the
call to `run` and then they try to use that handle aganin in the future.