Fixes #75 Made OutputLayout more ergonomic
The memory model has changed substantially with this patch.

OutputLayout no longer needs an Rc Refcell to use an Output or Cursor
with it. Instead, they keep internal OutputLayoutHandles to the
OutputLayout that is automatically cleaned up when the user drops the
OutputLayout.

If the Output dies before the OutputLayout, it removes itself from the
OutputLayout (as it did before).

If the OutputLayout dies first, then the weak handles are simply left as
they will not resolve.

If an Output is removed from the OutputLayout, their internal user data
is set to null (and the previous OutputLayoutHandle discarded).

If a Cursor dies before the OutputLayout, nothing happens as there's
nothing in the OutputLayout that depends on the Cursor.

There's still a safety hole with an OutputLayout dying before the cursor
though. That could lead to a dangling pointer. That will be fixed soon.