Updated wlroots to new version, fixed breakage
The only breakage was the move from resolution listener to mode listener
(e.g now the callback triggers whenever the mode changes at all, not
just the resolution.

The other change was the input to the seat functions, specifically for
the keyboard (they added modifiers and keycodes).

This neccesitated a change to KeyboardModifiers, it is now an alias to
the type returned by wlroots which is just a bag of data.