Added rationales and TODOS for safety
Main thing is having to check if an output is in the output layout.
Which we can work around by using OutputLayoutHandle's internally to the
cursor.

The main issue I have with that is now we'd have to keep them updated if
we allow moving cursors between outputs (which I forsee as something a
user would want to do)