Fix wl_shell_test so it works again w/ seat
Doesn't do anything, but it broadcasts the right thing to the client.

One thing that should be noted is that the setup for a Seat is very
awkward, requiring this weird Optional thing.

In a "real" compositor they could use a Vec to have multiple of them,
which would sorta solve the issue at least.

I'm ok with this for now, but it's more than a little awkward.

Note that the compositor write can get around this with clever use of
unsafe and ::std::mem::zeroed or whatever, though that is UB in the case
of a panic (which is entirely possible)