Totally changed Seat
It now uses wayland_listener! to set up the listeners and takse in the
dynamically dispatched handler which it uses for each of the events.

We directly return the Box-ed value that wayland_listener! returns,
which may be unsafe but it's such a small safety hole I'm not going to
worry about it...

We get around being able to access the Seat in both contexts by using a
RefCell

Fixed shell test