Added checks for null in destroy callbacks
Fixes #93

We forgot to handle the case where the compositor is shutting down and
we set the compositor ptr to null.

We have to set the ptr to null and not modify anything due to how
unwinding works. The struct will be in an indeterminate state when we
are in the destroy callback during a shutdown, so we shouldn't be donig
_anything_