Switch to hashmap based system for Seat.
Compositor writer no longer owns the Seat directly, but must access it
from a hashmap.

To access it, they need to use the name it's associated with (NOTE
Currently they can be desynched with the set_name on Seat. That will be
fixed).

To remove it, they need to call the Compositor::drop_seat method.

Once in a callback, you cannot access the seat because it has been
"moved from the hashmap"