Added cursor event callbacks.
Cursor event callbacks are now mandatory to state, using the same
mechanism one uses to define other callbacks.

The only thing that is currently not 100% great with this patch is that
it doesn't stop you from dropping or otherwise getting double mutable
access to the cursor during the callback. That will be fixed in the next
patch.