Added top level and v6 to xdgv6
Might need to tweak this though, e.g might need to remove the top level
and make the enum present at all levels.

Most of the functions associated with an xdg v6 shell surface has to do
with the top level, and actually error out otherwise. Thus to make sure
it's correct we should have the XdgV6ShellSurface actually be an enum
between the two variants that must be matched, with only the top level
having the desired functions and pinning the get functionality to each
enum variant.

Finally, the top-level getters can be on the enum type and each enum
variant can have their own version of the pointer to increase code
sharing (and Rust should be clever enough to remove the branching)