Made SubsurfaceManager, Surface::from_ptr -> new
With the change to Surface::new, found a bug in the wl shell destruction
event where we were passing an entirely new Surface to the destruction
callback, which could have resulted in multiple mutable borrows. Fixed
by turning it into a handle and then borrowing it for the call.

By putting the subshell management in a dedicated manager, we solved all
of the dangling pointer issues _and_ compartmentalized the logic.

The only thing left to solve now is proper cleanup of the signals in the
event of subsurface destruction (destruction callback) and in the case
of surface destruction (e.g the destructor of Surface)