Stubbed out input stuff
I realized I need to allow seats to have their own state and that the
cursors don't need to be stored in OutputLayouts (but I still need to
have restrictions on them).

Going to tangent now and work on that since that affects the rootston
API design in a big way.