Added with_compositor, which was pretty hard
Adding with_compositor (which allows you to use the compositor outside
of the wayland event loop) was rather hard because it broke a lot of
assumptions made that we had unique access to it everywhere.

So we added checks similar to the handles to ensure we do.

These series of patches also fixes a critical error made in cursor,
seat, and output layout handles in that you could upgrade them within a
callback even if they were mutably borrowed. That has since been fixed,
though that also took some effort.