Fixed the safety hole with textures
You can also destroy them now. This helps fix a bad memory leak in Way
Cooler (and any other WMs that create their own textures).

The memory model is a little weird, but it's required due to contraints
on when you can free the texture.

Specifically you can't free the texture when the renderer has begun
drawing, so it needs to be defined on the GenericRenderer.

Finally, you can only free ones you've created. In order to only have
one type, you can specify a type owned by the library user by creating
it and having the 'static lifetime associated with the struct. Only
instances with this lifetime can be freed (as otherwise it's handled
internally by wlroots, e.g. with surfaces from shell clients)