XWayland: fix segfault on new surface
On the new_surface event, the xwayland shell surface does not have a
cooresponding wlr_surface so trying to access it will segfault.

XWayland shell surfaces gain a new surface on each map event, so accept
the SurfaceHandler on that event instead of when the shell surface is
created.

Update some places in the code that assume the XWayland shell surface
will always have a wlr surface.