Put everything into nested modules
This is much easier to read the import list and honestly a lot better in
general.

We did have to make the public facing part of the library not match up
with the internal layout, which is fine honestly.

I do want to rename the event types to be less verbose, as well as the
other types. Really there's no need we can't do keyboard::Handle for
example.