Point out cursor and output::layout special cases
So it looks like our utils::Handle implementation isn't as generic as
we'd like it to be. Specifically it doesn't handle the "we own it, but
we need to have proxies to it for consistency" case that well due to the
need to move the value out during the upgrade.

Though perhaps we can get around that by making upgrade return a &mut
T, I don't think it's worth it to try to abstract it further.