Only one compositor::Builder per process
This fixes a glaring safety hole where you can reset the global
pointers, either while the compositor is running, after it was
constructed, or even from another thread.

This also removes the possibility of a data race for the global function
pointers in each resource manager. Before it was possible, now it is
not.