Compositor writer manually sets WAYLAND_DISPLAY
This is done to increase flexibility for three reasons:
1) If you want to spawn programs at startup you would need to set this
already, and setting it twice seems silly. So lets make it easier and
let that happen.
2) You may want to deny clients from connecting, though this is less a
security measure and more of a "lets lock this down or maybe forward it
to another compositor" use case. It's a small edge caset tbf.
3) It's trivial to do, so you don't lose much by having to do it
yourself.

Later on I might also make starting the backend a manual thing you must
do, but I need to ensure it's safe (the problem is I keep the Compositor
in a global `UnsafeCell` once I start the event loop, and I don't want
it to be possible to call the backedn start more than once so I'll need
to make it an affine type which is more rewriting than I'm willing to do
right now. Also we should probably make it return a result so the
backend failure can be dealt with)