Do use-after-free resource check better
Now we return an Option so we have to check it. This introduces overhead
that will be removed by switching to NonNull.

The one downside is one ugly, ugly thing:

Some(OptionalShellState(match self.state {
      None => None,
       Some(ref state) => Some(unsafe { state.clone() })
}))

This is a Some of an optional wrapper...so yeah it's really inefficient
and ugly in order to make it so that users can't clone shell states.